extraccion/requirements.txt
pyspark==3.2.1
requests==2.26.0
python-dotenv==0.19.2
pandas==1.3.3